requirements.txt
djangorestframework==3.15.2
psycopg2-binary==2.9.10
django-filter==25.1